chcpyth/requirements/common.txt
#Common requirements
Django==1.6.9
MySQL-Python==1.2.4
South==1.0
Pillow==2.8.1
django-tastypie==0.11.1
django-grappelli==2.5.4
django-extensions==1.4.9
python-Levenshtein
fuzzywuzzy
